eliminate some warning messages from C++
git-svn-id: svn://svn.macaulay2.com/Macaulay2/trunk@7613 a8012f80-5f13-0410-8dc0-eb77bd940c0d

+++ M2/Macaulay2/e/overflow.hpp
@@ -137,7 +137,7 @@ namespace safe {
 
      static inline int32_t mult(int32_t x, int32_t y, const char *msg) {
 	  int64_t z = (int64_t)x * y;
-	  int32_t w = z;
+	  int32_t w = (int32_t)z;
 	  if expect_false (z != (int64_t)w) ov(msg);
 	  return w;
      }
@@ -147,7 +147,7 @@ namespace safe {
 
      static inline int32_t mult_by(int32_t &x, int32_t y, const char *msg) {
 	  int64_t z = (int64_t)x * y;
-	  int32_t w = z;
+	  int32_t w = (int32_t)z;
 	  if expect_false (z != (int64_t)w) ov(msg);
 	  return x=w;
      }
